refactor directory hierarchy
