+++ src/test/java/com/revlearn/team1/unit/service/DiscussionPostServiceTest.java
@@ -1,4 +1,4 @@
-package com.revlearn.team1;
+package com.revlearn.team1.unit.service;
 
 
 import java.time.LocalDateTime;
